Improve Url helper so we can get rid of onBuildUrl
!!! This is not the final commit of this branch. There is still minor work to be done. !!!

The onBuildUrl event handler as it was implemented in the SiteHelper was
alway interfering with the creation of new OntoWiki_Url instances. This
especially caused trouble, when calling a specific controller/action from
a site (e.g. to use an extension).

With these improvements the according functionality is moved from the
SiteHelper to the Site_View_Helper_Url. With the code in this place one
can better control the behavior of created URLs. E.g. the new option
'stayOnSite' will tell the helper, whether the created url should keep
the user on the site or lead him to the OntoWiki backends properties view.

Furthermore the url helper is now also called using an options array (as
many other helpers).